requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
argilla==1.6.0
asgiref==3.6.0
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
blis==0.7.9
browse==1.0.1
bs4==0.0.1
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.1.0
chromadb==0.3.21
click==8.1.3
clickhouse-connect==0.5.22
cohere==4.3.0
colorama==0.4.6
comm==0.1.3
commonmark==0.9.1
confection==0.0.4
cryptography==40.0.1
cymem==2.0.7
dataclasses-json==0.5.7
datasets==2.12.0
debugpy==1.6.7
decorator==5.1.1
Deprecated==1.2.13
dill==0.3.6
Django==4.2
django-cors-headers==3.14.0
django-debug-toolbar==4.0.0
django-redis==5.2.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
duckdb==0.7.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==1.2.0
faiss-cpu==1.7.4
fastapi==0.95.1
filelock==3.12.0
frozenlist==1.3.3
fsspec==2023.4.0
google-api-core==2.11.0
google-api-python-client==2.85.0
google-auth==2.17.2
google-auth-httplib2==0.1.0
google-search-results==2.4.2
googleapis-common-protos==1.59.0
greenlet==2.0.2
h11==0.14.0
hnswlib==0.7.0
httpcore==0.16.3
httplib2==0.22.0
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.14.1
idna==3.4
ipykernel==6.22.0
ipython==8.12.0
jaraco.context==4.3.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
keyboard==0.13.5
langchain==0.0.150
langcodes==3.3.0
loguru==0.7.0
lxml==4.9.2
lz4==4.3.2
Markdown==3.4.3
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
monotonic==1.6
more-itertools==9.1.0
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
multiprocess==0.70.14
murmurhash==1.0.9
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.23.5
olefile==0.46
openai==0.27.4
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
outcome==1.2.0
packaging==23.0
pandas==1.5.3
parso==0.8.3
pathy==0.10.1
pickleshare==0.7.5
Pillow==9.5.0
pinecone-client==2.2.1
platformdirs==3.3.0
posthog==3.0.1
preshed==3.0.8
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==5.9.5
psycopg2==2.9.6
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.7
Pygments==2.15.1
PyJWT==2.6.0
PyMuPDF==1.22.1
pypandoc==1.11
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-pptx==0.6.21
pytz==2023.3
pywin32==306
PyYAML==6.0
pyzmq==25.0.2
redis==4.5.4
regex==2023.3.23
requests==2.28.2
responses==0.18.0
rfc3986==1.5.0
rich==13.0.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
selenium==4.9.0
sentence-transformers==2.2.2
sentencepiece==0.1.98
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==1.4.47
sqlparse==0.4.3
srsly==2.4.6
stack-data==0.6.2
starlette==0.26.1
sympy==1.11.1
tenacity==8.2.2
termcolor==2.3.0
thinc==8.1.9
threadpoolctl==3.1.0
tiktoken==0.3.3
tokenizers==0.13.3
torch==2.0.0
torchvision==0.15.1
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
transformers==4.28.1
trio==0.22.0
trio-websocket==0.10.2
typer==0.7.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
unstructured==0.6.1
uritemplate==4.1.1
urllib3==1.26.15
uuid==1.30
uvicorn==0.22.0
wasabi==1.1.1
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.2
win32-setctime==1.1.0
wolframalpha==5.0.0
wrapt==1.14.1
wsproto==1.2.0
XlsxWriter==3.1.0
xmltodict==0.13.0
xxhash==3.2.0
yarl==1.8.2
zstandard==0.21.0
